moved fuel-analysis component
